requirements.txt
Django>=3.0
PyMySQL~=1.0.2

Pillow~=8.2.0
requests~=2.25.1


mux_python
pip install git+https://github.com/muxinc/mux-python.git